Original TextEdit source from 10.6 SDK

+++ MultiplePageView.h
@@ -4,17 +4,17 @@
     NSPrintInfo *printInfo;
     NSColor *lineColor;
     NSColor *marginColor;
-    unsigned numPages;
+    NSUInteger numPages;
 }
 
 - (void)setPrintInfo:(NSPrintInfo *)anObject;
 - (NSPrintInfo *)printInfo;
-- (float)pageSeparatorHeight;
+- (CGFloat)pageSeparatorHeight;
 - (NSSize)documentSizeInPage;	/* Returns the area where the document can draw */
-- (NSRect)documentRectForPageNumber:(unsigned)pageNumber;	/* First page is page 0 */
-- (NSRect)pageRectForPageNumber:(unsigned)pageNumber;	/* First page is page 0 */
-- (void)setNumberOfPages:(unsigned)num;
-- (unsigned)numberOfPages;
+- (NSRect)documentRectForPageNumber:(NSUInteger)pageNumber;	/* First page is page 0 */
+- (NSRect)pageRectForPageNumber:(NSUInteger)pageNumber;	/* First page is page 0 */
+- (void)setNumberOfPages:(NSUInteger)num;
+- (NSUInteger)numberOfPages;
 - (void)setLineColor:(NSColor *)color;
 - (NSColor *)lineColor;
 - (void)setMarginColor:(NSColor *)color;
